ECO-56: AuthService interface.
